Refactor: warn and prepare functions are placed in the shared module
