+++ lib/runners.rb
@@ -65,6 +65,7 @@ require "runners/python"
 require "runners/swift"
 require "runners/go"
 require "runners/cplusplus"
+require "runners/recommended_config"
 require "runners/schema/options"
 require "runners/schema/trace"
 require "runners/schema/config"
