+++ sig/runners/recommended_config.rbi
@@ -1,0 +1,4 @@
+module Runners::RecommendedConfig : Processor
+  def warn_recommended_config_file_release: () -> Any
+  def prepare_config_file: () -> Any
+end
